# frontend/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/articles": {
      "get": {
        "summary": "page",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIPageArticleListVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "title": "",
                        "summary": "",
                        "coverUrl": "",
                        "type": "",
                        "price": 0,
                        "status": "",
                        "publishTime": "",
                        "viewCount": 0,
                        "likeCount": 0,
                        "pinned": false,
                        "pinnedTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "创建文章：支持“保存草稿”与“直接发布”两种行为。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreateDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLong"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": 0
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/articles/{id}": {
      "get": {
        "summary": "detail",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "draft",
            "in": "query",
            "description": "",
            "required": true,
            "example": "false",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseArticleDetailVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "authorId": 0,
                    "title": "",
                    "summary": "",
                    "coverUrl": "",
                    "categoryId": 0,
                    "type": "",
                    "price": 0,
                    "status": "",
                    "allowComment": false,
                    "publishTime": "",
                    "contentMd": "",
                    "contentHtml": "",
                    "attachments": "",
                    "viewCount": 0,
                    "likeCount": 0,
                    "commentCount": 0,
                    "locked": false,
                    "purchased": false,
                    "canManageAttachments": false,
                    "owner": false,
                    "pinned": false,
                    "pinnedTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdateDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "delete",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/articles/hot": {
      "get": {
        "summary": "hot",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListArticleListVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "title": "",
                      "summary": "",
                      "coverUrl": "",
                      "type": "",
                      "price": 0,
                      "status": "",
                      "publishTime": "",
                      "viewCount": 0,
                      "likeCount": 0,
                      "pinned": false,
                      "pinnedTime": ""
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/articles/{id}/publish": {
      "post": {
        "summary": "publish",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/articles/{id}/view": {
      "post": {
        "summary": "increaseView",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/articles/{id}/like": {
      "post": {
        "summary": "like",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/articles/{id}/publish": {
      "post": {
        "summary": "publish",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/ai/assistant": {
      "post": {
        "summary": "process",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAiResponseDTO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "scene": "",
                    "prompt": "",
                    "response": "",
                    "tokenUsage": 0,
                    "createTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/ai/assistant/history": {
      "get": {
        "summary": "history",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAiResponseDTO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "scene": "",
                      "prompt": "",
                      "response": "",
                      "tokenUsage": 0,
                      "createTime": ""
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseLong": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "integer",
            "description": "具体数据内容",
            "format": "int64"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "description": "具体数据内容",
            "type": "null"
          }
        }
      },
...
      "ArticleListVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": ""
          },
          "summary": {
            "type": "string",
            "description": ""
          },
          "coverUrl": {
            "type": "string",
            "description": ""
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "publishTime": {
            "type": "string",
            "description": ""
          },
          "viewCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "likeCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "pinned": {
            "type": "boolean",
            "description": ""
          },
          "pinnedTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "IPageArticleListVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleListVO",
              "description": "文章列表项。"
            },
            "description": ""
          },
          "total": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": ""
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": ""
          },
          "searchCount": {
            "type": "boolean",
            "description": ""
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": ""
          },
          "maxLimit": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": ""
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseIPageArticleListVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/IPageArticleListVO",
            "description": "具体数据内容"
          }
        }
      },
      "ArticleDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "authorId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": ""
          },
          "summary": {
            "type": "string",
            "description": ""
          },
          "coverUrl": {
            "type": "string",
            "description": ""
          },
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "allowComment": {
            "type": "boolean",
            "description": ""
          },
          "publishTime": {
            "type": "string",
            "description": ""
          },
          "contentMd": {
            "type": "string",
            "description": ""
          },
          "contentHtml": {
            "type": "string",
            "description": ""
          },
          "attachments": {
            "type": "string",
            "description": ""
          },
          "viewCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "likeCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "locked": {
            "type": "boolean",
            "description": ""
          },
          "purchased": {
            "type": "boolean",
            "description": ""
          },
          "canManageAttachments": {
            "type": "boolean",
            "description": ""
          },
          "owner": {
            "type": "boolean",
            "description": ""
          },
          "pinned": {
            "type": "boolean",
            "description": ""
          },
          "pinnedTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseArticleDetailVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/ArticleDetailVO",
            "description": "具体数据内容"
          }
        }
      },
      "ApiResponseListArticleListVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleListVO",
              "description": "文章列表项。"
            },
            "description": "具体数据内容"
          }
        }
      },
      "ArticleCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "description": "",
            "maxLength": 512
          },
          "coverUrl": {
            "type": "string",
            "description": "",
            "maxLength": 255
          },
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "allowComment": {
            "type": "boolean",
            "description": ""
          },
          "contentMd": {
            "type": "string",
            "description": ""
          },
          "contentHtml": {
            "type": "string",
            "description": ""
          },
          "attachments": {
            "type": "string",
            "description": ""
          },
          "publish": {
            "type": "boolean",
            "description": "是否直接发布，默认保存为草稿"
          }
        },
        "required": [
          "title",
          "type",
          "price",
          "contentMd"
        ]
      },
      "ArticleUpdateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "",
            "maxLength": 255
          },
          "summary": {
            "type": "string",
            "description": "",
            "maxLength": 512
          },
          "coverUrl": {
            "type": "string",
            "description": "",
            "maxLength": 255
          },
          "categoryId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "allowComment": {
            "type": "boolean",
            "description": ""
          },
          "contentMd": {
            "type": "string",
            "description": ""
          },
          "contentHtml": {
            "type": "string",
            "description": ""
          },
          "attachments": {
            "type": "string",
            "description": ""
          },
          "publish": {
            "type": "boolean",
            "description": "是否发布，当为 false 时保存为草稿"
          }
        },
        "required": [
          "title",
          "contentMd"
        ]
      },
...
      "ApiResponseAiResponseDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/AiResponseDTO",
            "description": "具体数据内容"
          }
        }
      },
      "AiRequest": {
        "type": "object",
        "properties": {
          "scene": {
            "type": "string",
            "description": "",
            "pattern": "POLISH|EXTEND|SUMMARY|TRANSLATE|CONTINUE"
          },
          "content": {
            "type": "string",
            "description": "",
            "maxLength": 2000
          },
          "extra": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "scene",
          "content"
        ]
      },
      "ApiResponseListAiResponseDTO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiResponseDTO",
              "description": "com.bluemsun.blog.module.ai.dto.AiResponseDTO"
            },
            "description": "具体数据内容"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}